+++ packages/backend/package.json
@@ -15,7 +15,7 @@
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
     "cron": "^1.8.2",
-    "db-utils": "1.0.0",
+    "util-db": "1.0.0",
     "express": "^4.17.1",
     "express-async-errors": "^3.1.1",
     "helmet": "^3.22.0",
