+++ packages/backend/src/surveys/initialHousehold/submissionData.js
@@ -1,4 +1,4 @@
-const { mongoose } = require("db-utils");
+const { mongoose } = require("util-db");
 const { createModel } = require("../../utils/mongoose");
 const { getSubmissionMetadata, FormSchema } = require("../sharedDataSchemas");
 const {
