+++ packages/backend/src/volunteer/volunteerData.js
@@ -1,4 +1,4 @@
-const { mongoose } = require("db-utils");
+const { mongoose } = require("util-db");
 const _ = require("lodash");
 const { createModel } = require("../utils/mongoose");
