+++ packages/backend/test/auth/routes/getAuthRoute.test.js
@@ -1,6 +1,6 @@
 const { getApp } = require("../../../src/app");
 
-const db = require("db-utils/inMemoryDb");
+const db = require("util-db/inMemoryDb");
 const supertest = require("supertest");
 
 const { login } = require("../../utils/requests");
