+++ packages/backend/test/auth/routes/verifyTokenRoute.test.js
@@ -4,7 +4,7 @@ const { findCookiesByVolunteerEmail } = require("../../../src/auth/cookieData");
 
 const { getApp } = require("../../../src/app");
 
-const db = require("db-utils/inMemoryDb");
+const db = require("util-db/inMemoryDb");
 const supertest = require("supertest");
 const { TEST_VOLUNTEER } = require("../../utils/requests");
