+++ packages/backend/test/surveySubmissions/routes.test.js
@@ -1,6 +1,6 @@
 const { getApp } = require("../../src/app");
 
-const db = require("db-utils/inMemoryDb");
+const db = require("util-db/inMemoryDb");
 
 const { login } = require("../utils/requests");
