+++ packages/util-db/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "db-utils",
+  "name": "util-db",
   "version": "1.0.0",
   "dependencies": {
     "mongodb-memory-server": "^6.6.1",
